+++ jaxrs/src/main/java/de/taimos/dvalin/jaxrs/providers/JacksonProvider.java
@@ -9,9 +9,9 @@ package de.taimos.dvalin.jaxrs.providers;
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
- * 
+ *
  *      http://www.apache.org/licenses/LICENSE-2.0
- * 
+ *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -53,10 +53,7 @@ public class JacksonProvider implements MessageBodyReader<Object>, MessageBodyWr
 
     @Override
     public boolean isWriteable(Class<?> type, Type genericType, Annotation[] annotations, MediaType mediaType) {
-        if (mediaType.getSubtype().equals("json") || mediaType.getSubtype().endsWith("+json")) {
-            return true;
-        }
-        return false;
+        return mediaType.getSubtype().equals("json") || mediaType.getSubtype().endsWith("+json");
     }
 
     @Override
@@ -71,15 +68,12 @@ public class JacksonProvider implements MessageBodyReader<Object>, MessageBodyWr
 
     @Override
     public boolean isReadable(Class<?> type, Type genericType, Annotation[] annotations, MediaType mediaType) {
-        if (mediaType.getSubtype().equals("json") || mediaType.getSubtype().endsWith("+json")) {
-            return true;
-        }
-        return false;
+        return mediaType.getSubtype().equals("json") || mediaType.getSubtype().endsWith("+json");
     }
 
     @Override
     public Object readFrom(Class<Object> type, Type genericType, Annotation[] annotations, MediaType mediaType, MultivaluedMap<String, String> httpHeaders, InputStream entityStream) throws IOException, WebApplicationException {
-        if (genericType.equals(null)) {
+        if (genericType == null) {
             return this.mapper.readValue(entityStream, type);
         }
         return this.mapper.readValue(entityStream, TypeFactory.defaultInstance().constructType(genericType));